Modularize the parsing implementation
The parsers have been moved into the crate root as a library. Parsers
are now exposed using TryFrom so we don't need to surface misc nom APIs
(the error type is still exposed).

To use the parsers, use the TryFrom impl on ConsumerOffsetsMessageKey
and OffsetCommitValue and pass in a byte buffer.

The example binary has been updated to utilize that.